move histogram
