+++ runtime/src/lib.rs
@@ -26,6 +26,7 @@ use std::{
 use thiserror::Error;
 
 pub mod deterministic;
+pub mod metrics;
 pub mod mocks;
 cfg_if::cfg_if! {
     if #[cfg(not(target_arch = "wasm32"))] {
