+++ runtime/src/metrics/mod.rs
@@ -1,6 +1,4 @@
 //! Utility functions for metrics
 
-mod histogram;
-
+pub mod histogram;
 pub mod status;
-pub use histogram::HistogramExt;
