+++ utils/Cargo.toml
@@ -13,7 +13,6 @@ documentation = "https://docs.rs/commonware-utils"
 [dependencies]
 bytes = { workspace = true }
 futures = { workspace = true }
-prometheus-client = { workspace = true }
 prost = { workspace = true }
 thiserror = { workspace = true }
