+++ utils/src/lib.rs
@@ -9,7 +9,6 @@ pub use time::SystemTimeExt;
 mod priority_set;
 pub use priority_set::PrioritySet;
 pub mod futures;
-pub mod metrics;
 
 /// Converts bytes to a hexadecimal string.
 pub fn hex(bytes: &[u8]) -> String {
